+++ application/controllers/EquipmentsController.php
@@ -4,7 +4,13 @@
  
  class EquipmentsController extends Application
  {
- 
+	function __construct()
+	{
+        parent::__construct();
+        $this->load->model('assetscsv');
+        $this->data['pagetitle'] = 'Equipments build';
+        $this->data['pagebody'] = 'assembly';
+	}
  	/**
  	 * Index Page for this controller.
  	 *
@@ -19,9 +25,10 @@
  	 */
  	public function index()
  	{
+            /*
  		$this->data['pagetitle'] = 'Equipments build';
  		$this->data['pagebody'] = 'assembly';
- 
+            
  		$parts = array();
  
  		$source = $this->equipments->all();
@@ -31,5 +38,18 @@
  		}
  		$this->data['parts'] = $parts;
  		$this->render();
+             */
+        $this->data['category'] = " all-equipments";        
+        $items = $this->assetscsv->all();
+        $this->data['items'] = $items;
+        $this->render();
  	}
+        
+    public function category($key)
+    {
+        $this->data['category'] = " ".$key;        
+        $items = $this->assetscsv->some('category', $key);
+        $this->data['items'] = $items;
+        $this->render();
+    }
  }