+++ application/controllers/PresetsController.php
@@ -4,7 +4,11 @@
  
  class PresetsController extends Application
  {
- 
+    function __construct()
+	{
+        parent::__construct();
+        $this->load->model('presetscsv');
+	}
  	/**
  	 * Index Page for this controller.
  	 *
@@ -19,9 +23,9 @@
  	 */
  	public function index()
  	{
- 		$this->data['pagetitle'] = 'Equipments presets';
- 		$this->data['pagebody'] = 'presets';
- 
+ 		// $this->data['pagetitle'] = 'Equipments presets';
+ 		// $this->data['pagebody'] = 'presets';
+                /*
  		$parts = array();
  
  		$source = $this->presets->all();
@@ -31,5 +35,17 @@
  		}
  		$this->data['parts'] = $parts;
  		$this->render();
+                
+                */
+                
+        $parts = $this->presetscsv->all();
+        $this->data['parts'] = $parts;
+        $this->render();
  	}
+
+    public function loadPreset($postVal) {
+        $img = $this->presetscsv->some('name', $postVal);
+        $this->data['preset'] = $img;
+    }
+
  }