+++ application/views/assembly.php
@@ -1,11 +1,11 @@
 <div class="POE_assembly">
-    <h1 id="assembly-title">Assembly - Path of Exile custom mod</h1>
+    <h1 id="assembly-title">Equipment - Path of Exile Inventory Simulator</h1>
     <div class="assembly-content">
-        <p> This is the Assembly page.</p>
+        <p>This is the{category} page.</p>
     </div>
     <div class="row">
- 	{parts}
- 		<div class="col-xs-4"><img class="img-responsive" src="/assets/img/{equip_code}.png" title="{equip_id}"></div>
- 	{/parts}
+ 	{items}
+ 		<div class="col-xs-4"><img class="img-responsive" src="/assets/img/{name}.png" title="{name}"></div>
+ 	{/items}
     </div> 
 </div>