+++ application/views/template.php
@@ -9,6 +9,7 @@
         <link rel="stylesheet" type="text/css" href="/assets/css/bootstrap.css"/>
         <link rel="stylesheet" type="text/css" href="/assets/css/bootstrap.min.css"/>
         <script src="/assets/js/jquery-3.1.0.min.js" type="text/javascript"></script>
+        <script src="/assets/js/presetHandler.js" type="text/javascript"></script>y
         </head>
 	<body>
             <nav class="navbar navbar-default navbar-fixed-top">
@@ -21,9 +22,16 @@
                   <ul class="nav navbar-nav">
                     <li><a href="/homepage">Homepage</a></li>
                     <li><a href="">Equipments</a>
-                        <ul>
-                            <li><a href="/presetscontroller">Presets</a></li>
-                            <li><a href="/equipmentscontroller">Equipments</a></li>
+                        <ul class="dropdown-menu" role="menu">
+                            <li><a role="menuitem" href="/equipmentscontroller">All Equipments</a></li>
+                            <li><a role="menuitem" href="/category/amulet">Amulet</a></li>
+                            <li><a role="menuitem" href="/category/belt">Belt</a></li>
+                            <li><a role="menuitem" href="/category/boot">Boots</a></li>
+                            <li><a role="menuitem" href="/category/glove">Gloves</a></li>
+                            <li><a role="menuitem" href="/category/helmet">Helmet</a></li>
+                            <li><a role="menuitem" href="/category/chest">Chest</a></li>
+                            <li><a role="menuitem" href="/category/ring">Ring</a></li>
+                            <li><a role="menuitem" href="/category/weapon">Weapon</a></li>
                         </ul>
                     </li>
                     <li><a href="/history">History</a></li>